etl/requirements.txt
firebirdsql==1.3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyodbc==4.0.39
numpy==2.1.3
pandas==2.2.3
statsmodels==0.14.4
